Reversing a conditional statement used for debugging
The statement:
<field name="domain">[("islocaluom","=",True)]</field>

Was used to verify correct creation of per product uom's.  A normal uom should automatically be created and set with islocaluom=True.  This line would normally filter these entries out of the normal UoM menus, but for debugging I had hidden the normal UoMs and showed only the per product UoMs